Forth lesson [+] Code review
