--- main/java/com/bchupika/lesson3/Address.java
@@ -1,3 +1,0 @@
-package com.bchupika.lesson3;
-
-public record Address(String street, String city) { }
