--- main/java/com/bchupika/lesson3/Grade.java
@@ -1,4 +1,0 @@
-package com.bchupika.lesson3;
-
-public record Grade(String name, double score) {
-}
