--- main/java/com/bchupika/user/dto/UserResponseDto.java
@@ -1,4 +1,0 @@
-package com.bchupika.user.dto;
-
-public record UserResponseDto(Long id, String name, String email, String phone) {
-}
