--- main/java/com/bchupika/user/exception/DataProcessingException.java
@@ -1,9 +1,0 @@
-package com.bchupika.user.exception;
-
-import java.sql.SQLException;
-
-public class DataProcessingException extends RuntimeException {
-  public DataProcessingException(String message, SQLException e) {
-    super(message);
-  }
-}
