--- main/java/com/bchupika/user/exception/EmailException.java
@@ -1,7 +1,0 @@
-package com.bchupika.user.exception;
-
-public class EmailException extends RuntimeException {
-    public EmailException(String message) {
-        super(message);
-    }
-}
