--- main/java/com/bchupika/user/exception/PhoneNumberException.java
@@ -1,7 +1,0 @@
-package com.bchupika.user.exception;
-
-public class PhoneNumberException extends RuntimeException {
-    public PhoneNumberException(String message) {
-        super(message);
-    }
-}
