+++ main/java/com/learn/lesson3/Address.java
@@ -1,0 +1,3 @@
+package com.learn.lesson3;
+
+public record Address(String street, String city) { }
