+++ main/java/com/learn/lesson3/Grade.java
@@ -1,0 +1,4 @@
+package com.learn.lesson3;
+
+public record Grade(String name, double score) {
+}
