+++ main/java/com/learn/user/dto/UserRegistrationDto.java
@@ -1,0 +1,6 @@
+package com.learn.user.dto;
+
+public record UserRegistrationDto(String name, String lastName, String email, String phoneNumber, String password,
+                                  String repeatPassword) {
+
+}
