+++ main/java/com/learn/user/dto/UserResponseDto.java
@@ -1,0 +1,4 @@
+package com.learn.user.dto;
+
+public record UserResponseDto(Long id, String name, String email, String phone) {
+}
