+++ main/java/com/learn/user/exception/DataProcessingException.java
@@ -1,0 +1,9 @@
+package com.learn.user.exception;
+
+import java.sql.SQLException;
+
+public class DataProcessingException extends RuntimeException {
+  public DataProcessingException(String message, SQLException e) {
+    super(message);
+  }
+}
