+++ main/java/com/learn/user/exception/ValidationException.java
@@ -1,0 +1,7 @@
+package com.learn.user.exception;
+
+public class ValidationException extends RuntimeException {
+    public ValidationException(String message) {
+        super(message);
+    }
+}
